CheckConntionAvalible

+++ lib/type_define.h
@@ -109,7 +109,7 @@ typedef std::function<void(void)> func;
 #define HOUR				3600
 #define ONE_DAY				86400
 
-#define NowTime             uint32(time(0))
+#define NowTime             uint64(time(0))
 
 template <bool x> struct STATIC_ASSERTION_FAILURE;
 template <> struct STATIC_ASSERTION_FAILURE<true> { enum {value = 1};};
